rpa/kavana/pyproject.toml
[project]
name = "kavana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "easyocr>=1.7.2",
    "imagehash>=4.3.2",
    "keyboard>=0.13.5",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.12",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "paramiko>=3.5.1",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "psycopg2>=2.9.10",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.13.0",
    "pylint>=3.3.7",
    "pymysql>=1.1.1",
    "pytesseract>=0.3.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "pywin32>=310",
    "pywinauto>=0.6.9",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.3",
    "screeninfo>=0.8.1",
    "selenium>=4.31.0",
    "vulture>=2.14",
    "webdriver-manager>=4.0.2",
    "xlrd>=2.0.1",
    "xmltodict>=0.14.2",
]
